+++ core/src/main/scala/org/apache/spark/sql/crossdata/XDContext.scala
@@ -19,16 +19,21 @@
 
 package org.apache.spark.sql.crossdata
 
-
-import java.lang.reflect.Constructor
+import java.util.ServiceLoader
 import java.util.concurrent.atomic.AtomicReference
 
-import com.typesafe.config.{Config, ConfigFactory}
-import org.apache.spark.sql._
-import org.apache.spark.sql.catalyst.{CatalystConf, SimpleCatalystConf}
+import com.stratio.crossdata.connector.FunctionInventory
+import org.apache.spark.sql.catalyst.analysis.Analyzer
+import org.apache.spark.sql.execution.ExtractPythonUDFs
+import org.apache.spark.sql.crossdata.execution.{XDStrategies, NativeUDF, ExtractNativeUDFs}
+import org.apache.spark.sql.execution.datasources.{PreWriteCheck, PreInsertCastAndRename}
 import org.apache.spark.sql.sources.crossdata.XDDdlParser
+import org.apache.spark.sql.{DataFrame, SQLContext, Strategy}
+import org.apache.spark.util.Utils
 import org.apache.spark.{Logging, SparkContext}
-
+import org.apache.spark.sql.catalyst.{CatalystConf, SimpleCatalystConf}
+import com.typesafe.config.{Config, ConfigFactory}
+import java.lang.reflect.Constructor
 
 /**
  * CrossdataContext leverages the features of [[SQLContext]]
@@ -50,24 +55,47 @@ class XDContext(@transient val sc: SparkContext) extends SQLContext(sc) with Log
       new SimpleCatalystConf(caseSensitive), self).asInstanceOf[XDCatalog]
   }
 
-  override protected[sql] val ddlParser = new XDDdlParser(sqlParser.parse)
+  @transient
+  override protected[sql] lazy val analyzer: Analyzer =
+    new Analyzer(catalog, functionRegistry, conf) {
+      override val extendedResolutionRules =
+        ExtractPythonUDFs ::
+          ExtractNativeUDFs::
+          PreInsertCastAndRename ::
+          Nil
+
+      override val extendedCheckRules = Seq(
+        PreWriteCheck(catalog)
+      )
+    }
 
   @transient
-  override protected[sql] val planner = new SparkPlanner with XDStrategies {
-
-    override def strategies: Seq[Strategy] =
-      XDDDLStrategy +: super.strategies
+  class XDPlanner extends SparkPlanner with XDStrategies {
+    override def strategies: Seq[Strategy] = Seq(XDDDLStrategy, NativeUDFStrategy) ++ super.strategies
   }
+
+  @transient
+  override protected[sql] val planner: SparkPlanner = new XDPlanner
+
+  @transient
+  protected[sql] override val ddlParser = new XDDdlParser(sqlParser.parse(_))
 
   override def sql(sqlText: String): DataFrame = {
     XDDataFrame(this, parseSql(sqlText))
   }
 
+  { //Register built-in UDFs for each provider available.
+    import scala.collection.JavaConversions._
+    val loader = Utils.getContextOrSparkClassLoader
+    val serviceLoader = ServiceLoader.load(classOf[FunctionInventory], loader)
+    for(srv <- serviceLoader.iterator();
+        inventory = srv.getClass.newInstance();
+        udf <- srv.nativeBuiltinFunctions
+    ) functionRegistry.registerFunction(udf.name, e => NativeUDF(udf.name, udf.returnType, e))
+  }
+
   XDContext.setLastInstantiatedContext(self)
-
 }
-
-
 
 /**
  * This XDContext object contains utility functions to create a singleton XDContext instance,
@@ -107,5 +135,6 @@ object XDContext {
       lastInstantiatedContext.set(xdContext)
     }
   }
+
 }