+++ driver/src/test/scala/com/stratio/crossdata/driver/DriverIT.scala
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.stratio.crossdata.driver
 
 import java.nio.file.Paths
+import java.util.concurrent.TimeoutException
 
 import akka.util.Timeout
 import com.stratio.crossdata.common.SQLCommand
@@ -24,6 +26,7 @@ import org.apache.spark.sql.AnalysisException
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 
+import scala.concurrent.Await
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
@@ -42,7 +45,6 @@ class DriverIT extends EndToEndTest {
     result.asInstanceOf[ErrorResult].cause.get.getMessage should include regex "cannot resolve .*"
   }
 
-
   it should "return a SuccessfulQueryResult when executing a select *" in {
     assumeCrossdataUpAndRunning()
     val driver = Driver()
@@ -57,8 +59,7 @@ class DriverIT extends EndToEndTest {
     result shouldBe an[SuccessfulQueryResult]
     result.queryId should be (sqlCommand.queryId)
     result.hasError should be (false)
-    result.resultSet.isDefined should be (true)
-    val rows = result.resultSet.get
+    val rows = result.resultSet
     rows should have length 2
     rows(0) should have length 2
 
@@ -69,4 +70,22 @@ class DriverIT extends EndToEndTest {
     */
   }
 
+  "Crossdata driver" should "fail with a timeout when there is no server" in {
+    val driver = Driver()
+    val sqlCommand = SQLCommand("select * from any")
+    val result = driver.syncQuery(sqlCommand, Timeout(1 seconds), 1)
+    result.hasError should be (true)
+    a [RuntimeException] should be thrownBy result.resultSet
+    result.queryId should be (sqlCommand.queryId)
+    result shouldBe an [ErrorResult]
+    result.asInstanceOf[ErrorResult].message should include regex "(?i)timeout was exceed"
+
+  }
+
+
+  it should "return a future with a timeout when there is no server" in {
+    val driver = Driver()
+    val future = driver.asyncQuery(SQLCommand("select * from any"), Timeout(1 seconds), 1)
+    a [TimeoutException] should be thrownBy Await.result(future, 2 seconds)
+  }
 }