+++ CHANGELOG.md
@@ -5,3 +5,8 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 ## 1.0.0 (upcoming)  
 
 * Upgraded Spark 1.5.1
+* Mix data from different data stores. Cassandra, MongoDB and Elastic Search.
+* Native access to data stores.
+* Driver API for external projects like ODBC/JDBC.
+* Improved and extended SparkContext: XDContext.
+* P2P architecture: High availability and scalability.